Introduce nix-darwin
This is an attempt to configure the devices of engineers through Nix.
We have a few things that need to be installed and configured globally;
most importantly, the local domain for the new stack.
The setup is not trivial, so nix-darwin is a big help. It is much more
reliable that a bash script.