Make local domain module more robust (#4)
This commit introduces a set of changes to make the module "local
domain" more reliable. Specifically:

- The various launch daemons now depend on one another. Dnsmasq does not
start if the alias isn't set, and Nginx does not start if Dnsmasq is not
enabled. This solves an issue where some daemon would fail if started in
the wrong order.
- Nginx send its logs in a specific directory, which is created when the
service is started. This fixes the issue of Nginx exiting when the
default directory for logs (`/var/log/nginx`) does not exist.

Additionally, the various launch daemons have been renamed to fit
conventions, and the logs directory has been reorganized.